Ticket 01: fix Place.owner resolution and averageRating precision
Adds this codebase's first type-level GraphQL field resolver
(Place.owner, fetching the User by ownerId) - previously declared in
the schema but never resolved, silently returning null. Also migrates
providers_places.average_rating from INTEGER to DECIMAL(2,1) so a
computed average like 4.4 isn't truncated to 4, and updates the
GraphQL schema's averageRating field from Int to Float to match -
without that, the decimal value fails to serialize at all.

Both verified live: owner now resolves correctly and distinctly for
different places, and a 4.4 rating round-trips through GraphQL exactly
instead of erroring or truncating.

+++ src/models/places.ts
@@ -51,7 +51,7 @@ const Place = sequelize.define<PlaceModelInterface>(
             field: "category_id"
         },
         averageRating: {
-            type: Sequelize.INTEGER,
+            type: Sequelize.DECIMAL(2, 1),
             allowNull: true,
             defaultValue: 0,
             validate: {
